Inline core classes into main module
